+++ content/chapters/chapter-3/index.tsx
@@ -20,7 +20,16 @@ export const metadata = {
   image: '/assets/images/chapter-3-cover.jpg',
   theme: 'bg-gradient-to-tl from-[#3C1843] to-[#553360]',
   intros: ['intro-1', 'intro-2'],
-  lessons: ['solo-1', 'coop-1', 'pool-1', 'split-1'],
+  lessons: [
+    'solo-1',
+    'coop-1',
+    'coop-2',
+    'pool-1',
+    'pool-2',
+    'pool-3',
+    'split-1',
+    'split-2',
+  ],
   outros: ['outro-1'],
   challenges: ['solo-1', 'coop-1', 'pool-1', 'split-1'],
 }
