+++ content/lessons/chapter-3/index.ts
@@ -4,6 +4,7 @@ import * as Pool1 from './pool-1'
 import * as Pool2 from './pool-2'
 import * as Pool3 from './pool-3'
 import * as Split1 from './split-1'
+import * as Split2 from './split-2'
 import * as Solo1 from './solo-1'
 import * as Outro1 from './outro-1'
 import * as Coop1 from './coop-1'
@@ -17,6 +18,7 @@ const Chapter3 = {
   'pool-2': Pool2,
   'pool-3': Pool3,
   'split-1': Split1,
+  'split-2': Split2,
   'outro-1': Outro1,
   'coop-1': Coop1,
   'coop-2': Coop2,
