+++ lib/progress.ts
@@ -35,6 +35,7 @@ export const keys = [
   'CH3POL2',
   'CH3POL3',
   'CH3SPL1',
+  'CH3SPL2',
   'CH3OUT1',
 ]
 
@@ -73,6 +74,7 @@ export const keysMeta = {
   CH3POL2: { path: '/chapter-3/pool-2' },
   CH3POL3: { path: '/chapter-3/pool-3' },
   CH3SPL1: { path: '/chapter-3/split-1' },
+  CH3SPL2: { path: '/chapter-3/split-2' },
   CH3OUT1: { path: '/chapter-3/outro-1' },
 }
